Parameter Bound changes
In TireModelFittingMain, added Dt and Et in order to get intuitive sense of where parameters should be.

In PureAligningNominal, messing around with parameters to get better fit.

Same with PureAligningVariant, only messing around with qDz2, qEz1, qEz2, qEz3, etc but only D and E parameters

Nothing changed with the ResponseSurfaces